# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 苍穹外卖 v2 API
  version: 2.0.0
  summary: 苍穹外卖完全重写版接口契约（唯一真相）
  description: |
    # 苍穹外卖 v2

    本文件是 v2 的**接口唯一真相**,由 `docs/01-domain.md`(领域模型的唯一语义来源)推导而来。
    人读的通用约定、完整错误码表、接口清单索引、WebSocket 协议见 `docs/03-api.md`。

    ## 受众与路径
    - `/api/v1/admin/**` —— 员工(管理端 Web),令牌 `aud: admin`
    - `/api/v1/customer/**` —— 顾客(小程序),令牌 `aud: customer`
    - `/api/v1/notify/**` —— 支付平台回调,不使用 Bearer,走平台签名/解密校验

    ## 认证
    `Authorization: Bearer <accessToken>`。access token 有效期 2 小时,refresh token 7 天且可撤销(Redis)。
    员工令牌与顾客令牌**互不通用**:载荷 `aud` 声明受众,主体 id 放在 `sub`,角色放在 `role`。

    ## 响应约定
    - 成功:直接返回资源本身或数组,**不套 `{code,data}` 信封**;无返回体的写操作返回 `204 No Content`。
    - 分页列表:`{"records": [...], "page": 1, "pageSize": 20, "total": 137}`。
    - 失败:**正确的 HTTP 状态码** + 统一错误体 `{code, message, details?, traceId}`。
      - 400 参数校验失败 / 401 未认证或令牌过期 / 403 已认证但无权限
      - 404 资源不存在 / 409 唯一键冲突 / 422 业务规则不满足 / 500 未预期错误

    ## 数据约定
    - 金额:整数,**单位分**,字段名以 `Cents` 结尾(如 `payAmountCents`)。
    - 时间:ISO-8601 带偏移(如 `2025-01-01T12:00:00+08:00`);日期参数 `YYYY-MM-DD`。
    - JSON 字段 camelCase;分页参数统一 `?page=1&pageSize=20`(`pageSize` 上限 100,**超限钳制为 100**)。
    - 列表排序统一 `?sort=placedAt,desc`,示例值必须在后端白名单内。
  contact:
    name: 苍穹外卖 v2
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: 本地开发
  - url: /
    description: 同源部署(前端经网关/反向代理)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/auth/login:
    post:
      tags: [Identity-Employee]
      operationId: employeeLogin
      summary: 员工登录
      description: |
        使用用户名与密码登录,成功返回 access + refresh 令牌对。
        密码使用 BCrypt 校验;`status=0`(禁用)的员工登录返回 403 `EMPLOYEE_DISABLED`。
        登录成功写入 `last_login_at`。失败提示不区分"用户不存在"与"密码错误",一律 `AUTH_BAD_CREDENTIALS`。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: AUTH_BAD_CREDENTIALS
                message: 用户名或密码错误
                traceId: 0f3a1c2b4d5e6f70
        '403':
          description: 员工已被禁用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/admin/auth/logout:
    post:
      tags: [Identity-Employee]
      operationId: employeeLogout
      summary: 员工登出
      description: 撤销当前 refresh token(从 Redis 删除),access token 依靠短有效期自然失效。幂等:重复登出也返回 204。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: 已登出
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/auth/refresh:
    post:
      tags: [Identity-Employee]
      operationId: employeeRefreshToken
      summary: 员工刷新令牌
      description: |
        用 refresh token 换新的令牌对。refresh 记录存于 Redis,可撤销;
        刷新后旧 refresh token 立即失效(旋转 refresh token)。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/auth/password:
    put:
      tags: [Identity-Employee]
      operationId: changeEmployeePassword
      summary: 员工修改自己密码
      description: 修改当前登录员工的密码。旧密码错误返回 401 `AUTH_BAD_CREDENTIALS`;成功后所有 refresh token 被撤销,需重新登录。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: 修改成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/admin/auth/me:
    get:
      tags: [Identity-Employee]
      operationId: getCurrentEmployee
      summary: 查询当前登录员工
      description: 返回 access token 主体对应的员工资料(不含密码散列)。
      responses:
        '200':
          description: 当前员工
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/employees:
    get:
      tags: [Identity-Employee]
      operationId: pageEmployees
      summary: 员工分页查询
      description: 仅 ADMIN。可按用户名/姓名模糊、状态、角色筛选。`sort` 白名单:`createdAt`、`username`、`lastLoginAt`(默认 `createdAt,desc`)。
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - name: name
          in: query
          description: 用户名或姓名,模糊匹配
          schema:
            type: string
            maxLength: 32
            example: 张
        - name: status
          in: query
          description: 启用状态;不传表示全部
          schema:
            $ref: '#/components/schemas/EmployeeStatus'
        - name: role
          in: query
          description: 角色;不传表示全部
          schema:
            $ref: '#/components/schemas/EmployeeRole'
      responses:
        '200':
          description: 员工分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEmployee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Identity-Employee]
      operationId: createEmployee
      summary: 新增员工
      description: 仅 ADMIN。用户名全局唯一,冲突返回 409 `EMPLOYEE_USERNAME_TAKEN`。初始密码由请求给定。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/employees/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Identity-Employee]
      operationId: getEmployeeById
      summary: 按 id 查询员工
      description: 仅 ADMIN。
      responses:
        '200':
          description: 员工详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Identity-Employee]
      operationId: updateEmployee
      summary: 编辑员工
      description: |
        仅 ADMIN。可改姓名、手机号、角色、状态;用户名不可在本接口修改。
        ADMIN 不能把自己的角色降为 STAFF,也不能禁用自己(422 `EMPLOYEE_SELF_ROLE_CHANGE`, `EMPLOYEE_SELF_DISABLE`)。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: 更新后的员工
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/admin/employees/{id}/status:
    parameters:
      - $ref: '#/components/parameters/PathId'
    patch:
      tags: [Identity-Employee]
      operationId: changeEmployeeStatus
      summary: 启用/禁用员工
      description: 仅 ADMIN。禁用后该员工的 access/refresh 令牌立即失效(令牌校验时回查状态)。ADMIN 不能禁用自己。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusPatchRequest'
      responses:
        '204':
          description: 已更新
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  ##########################################################################
  # Identity - Customer
  ##########################################################################
  /api/v1/customer/auth/wechat-login:
    post:
      tags: [Identity-Customer]
      operationId: customerWechatLogin
      summary: 顾客微信登录
      description: |
        用微信 `code` 换取 `openid`(服务端调用 `code2Session`)。
        首次登录自动建档(昵称默认"微信用户")。令牌 `aud: customer`;`status=0`(封禁)返回 403 `CUSTOMER_DISABLED`。
        同一 openid 并发首登返回 409 `CUSTOMER_DUPLICATE_OPENID`,客户端重试即可。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerWechatLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: 顾客已被封禁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: 微信 code 已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: CUSTOMER_WECHAT_CODE_INVALID
                message: 微信登录凭证已失效,请重试
                traceId: 0f3a1c2b4d5e6f70
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/customer/auth/refresh:
    post:
      tags: [Identity-Customer]
      operationId: customerRefreshToken
      summary: 顾客刷新令牌
      description: 用 refresh token 换新令牌对,并旋转 refresh token。顾客令牌不能用于员工接口(`AUTH_AUDIENCE_MISMATCH`)。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customer/auth/logout:
    post:
      tags: [Identity-Customer]
      operationId: customerLogout
      summary: 顾客登出
      description: 撤销当前 refresh token。幂等:重复登出也返回 204。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: 已登出
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <accessToken>`。
        员工令牌 `aud=admin`、顾客令牌 `aud=customer`,受众不匹配返回 403 `AUTH_AUDIENCE_MISMATCH`。
        access token 有效期 2 小时;过期返回 401 `AUTH_TOKEN_EXPIRED`。
        `/api/v1/notify/**` 不使用本方案,走支付平台签名/解密校验。
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: 资源 id
      schema:
        type: integer
        format: int64
      example: 1
...
    Page:
      name: page
      in: query
      description: 页码,从 1 开始;小于 1 时按 1 处理
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSize:
      name: pageSize
      in: query
      description: |
        每页条数,默认 20,上限 100。**超限钳制为 100**(不报错),
        传入非正整数返回 400 `COMMON_VALIDATION_FAILED`。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    Sort:
      name: sort
      in: query
      description: |
        排序,格式 `字段,asc|desc`,例如 `sort=placedAt,desc`。
        字段必须在**该接口白名单**内,否则 400 `COMMON_SORT_FIELD_NOT_ALLOWED`;缺省 `direction` 视为 `asc`。
      schema:
        type: string
        pattern: '^[a-zA-Z][a-zA-Z0-9]*(,(asc|desc))?$'
      example: placedAt,desc
...
  responses:
    BadRequest:
      description: 参数校验失败(400 `COMMON_VALIDATION_FAILED`)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: COMMON_VALIDATION_FAILED
            message: 请求参数校验失败
            details:
              - field: pageSize
                reason: 必须为正整数
            traceId: 0f3a1c2b4d5e6f70
    Unauthorized:
      description: 未认证或令牌已过期(401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTH_TOKEN_EXPIRED
            message: 登录状态已过期,请重新登录
            traceId: 0f3a1c2b4d5e6f70
    Forbidden:
      description: 已认证但无权限(403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTH_AUDIENCE_MISMATCH
            message: 令牌受众不匹配,无法访问该接口
            traceId: 0f3a1c2b4d5e6f70
    NotFound:
      description: 资源不存在(404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: COMMON_RESOURCE_NOT_FOUND
            message: 请求的资源不存在
            traceId: 0f3a1c2b4d5e6f70
    Conflict:
      description: 唯一键冲突(409)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CATEGORY_NAME_TAKEN
            message: 同类型下已存在同名分类
            details:
              - field: name
                reason: 该名称已被占用
            traceId: 0f3a1c2b4d5e6f70
    UnprocessableEntity:
      description: 业务规则不满足(422)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ORDER_SHOP_CLOSED
            message: 门店已打烊,暂时无法下单
            traceId: 0f3a1c2b4d5e6f70
    InternalError:
      description: 未预期错误(500 `COMMON_INTERNAL_ERROR`)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: COMMON_INTERNAL_ERROR
            message: 服务器开小差了,请稍后重试
            traceId: 0f3a1c2b4d5e6f70
...
  schemas:
...
    ErrorDetail:
      type: object
      description: 单条字段级错误说明
      required: [reason]
      properties:
        field:
          type: string
          description: 出错的字段名(路径式,如 `flavors[0].name`);非字段级错误可省略
          maxLength: 128
          example: status
        reason:
          type: string
          description: 面向开发者的中文说明
          maxLength: 255
          example: DELIVERING 不能取消
    ErrorResponse:
      type: object
      description: |
        统一错误体。**与成功响应不同构**——成功响应直接是资源本身,绝不套壳。
        每个错误响应都带 `X-Trace-Id` 响应头(与 `traceId` 相同),便于串联日志。
      required: [code, message, traceId]
      properties:
        code:
          type: string
          description: 错误码,`SCREAMING_SNAKE_CASE`,按上下文前缀;完整清单见 `docs/03-api.md` 第 2.2 节
          pattern: '^[A-Z][A-Z0-9_]*$'
          example: ORDER_INVALID_TRANSITION
        message:
          type: string
          description: 面向用户的中文提示,前端可直接展示
          maxLength: 255
          example: 订单当前状态不允许该操作
        details:
          type: array
          description: 可选的字段级细节,主要用于 400/422
          items:
            $ref: '#/components/schemas/ErrorDetail'
        traceId:
          type: string
          description: 链路追踪 id,报障时请提供
          maxLength: 64
          example: 0f3a1c2b4d5e6f70
...
    EnabledFlag:
      type: integer
      description: 启用标志。分类/菜品/套餐:`1` 启用(起售)/ `0` 禁用(停售)
      enum: [1, 0]
      example: 1
    EmployeeStatus:
      type: integer
      description: 员工状态。`1` 启用 / `0` 禁用
      enum: [1, 0]
      example: 1
...
    EmployeeRole:
      type: string
      description: 员工角色。`ADMIN` 可管理员工,`STAFF` 不可(D3)
      enum: [ADMIN, STAFF]
      example: STAFF
    Employee:
      type: object
      description: 员工(不含密码散列)
      required: [id, username, name, role, status]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          description: 登录名,全局唯一
          maxLength: 32
          example: zhangsan
        name:
          type: string
          description: 姓名
          maxLength: 32
          example: 张三
        phone:
          type: string
          description: 手机号
          maxLength: 20
          example: '13800138000'
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: 最近登录时间,从未登录为 null
          example: '2025-01-01T12:00:00+08:00'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedBy:
          type: integer
          format: int64
    EmployeeLoginRequest:
      type: object
      description: 员工登录请求
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 32
          example: admin
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 64
          description: 明文密码,HTTPS 传输;服务端按 BCrypt 校验
          example: admin123
    EmployeeCreateRequest:
      type: object
      description: 新增员工请求;用户名全局唯一,初始密码由请求给定
      required: [username, password, name, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z][a-zA-Z0-9_]{2,31}$'
          description: 登录名,全局唯一
          example: lisi
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 64
          example: lisi1234
        name:
          type: string
          minLength: 1
          maxLength: 32
          example: 李四
        phone:
          type: string
          maxLength: 20
          pattern: '^1[3-9][0-9]{9}$'
          example: '13900139000'
        role:
          $ref: '#/components/schemas/EmployeeRole'
    EmployeeUpdateRequest:
      type: object
      description: 编辑员工。用户名不可改。
      required: [name, role, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
          example: 李四
        phone:
          type: string
          maxLength: 20
          pattern: '^1[3-9][0-9]{9}$'
          example: '13900139000'
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
    ChangePasswordRequest:
      type: object
      description: 修改当前登录员工密码
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
          format: password
          minLength: 6
          maxLength: 64
          example: admin123
        newPassword:
          type: string
          format: password
          minLength: 6
          maxLength: 64
          description: 不能与旧密码相同
          example: admin1234
    RefreshTokenRequest:
      type: object
      description: 刷新/登出请求;两种接口共用同一请求体
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 16
          description: 登录/刷新时下发的 refresh token(7 天有效,可撤销,存 Redis)
          example: rt_7f3c9a1e5b2d4086
    TokenPair:
      type: object
      description: |
        令牌对。员工令牌 `aud=admin`,顾客令牌 `aud=customer`,互不通用。
        载荷声明:`aud`(受众)、`sub`(主体 id)、`role`(员工为 ADMIN/STAFF,顾客为 CUSTOMER)。
      required: [accessToken, refreshToken, expiresIn, refreshExpiresIn, tokenType]
      properties:
        accessToken:
          type: string
          description: 访问令牌,有效期 2 小时,请求时放入 `Authorization` 头,Bearer 方案
          example: eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJhZG1pbiJ9.x
        refreshToken:
          type: string
          description: 刷新令牌,有效期 7 天,刷新后旋转(旧值立即失效)
          example: rt_7f3c9a1e5b2d4086
        expiresIn:
          type: integer
          description: access token 剩余有效期(秒),固定 7200
          example: 7200
        refreshExpiresIn:
          type: integer
          description: refresh token 剩余有效期(秒),固定 604800
          example: 604800
        tokenType:
          type: string
          description: 固定为 `Bearer`
          enum: [Bearer]
          example: Bearer
...
    CustomerWechatLoginRequest:
      type: object
      description: 顾客微信登录请求;`code` 单次有效
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 128
          description: 小程序 `wx.login` 返回的临时登录凭证,服务端用它换 `openid`
          example: 081Kq7Ga1xY2FG0zJQHa1ZPbGa1Kq7Ga
        nickname:
          type: string
          maxLength: 64
          description: 可选。首次登录建档时的昵称,不传则用"微信用户"
          example: 小明
        avatarUrl:
          type: string
          maxLength: 255
          description: 可选。首次登录建档时的头像 URL
          example: https://cdn.example.com/avatar/new.png
...
    StatusPatchRequest:
      type: object
      description: 单资源启用/禁用
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/EnabledFlag'
...
    PagedEmployee:
      type: object
      description: 员工分页
      required: [records, page, pageSize, total]
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 137
...